Develop_BJH

+++ yagaza/src/main/java/com/spring/yagaza/web/service/FileService.java
@@ -6,6 +6,6 @@ import org.springframework.web.multipart.MultipartFile;
 public interface FileService {
     ImgFile store(MultipartFile file) throws Exception;
 
-    ImgFile load(int id) throws Exception;
+    ImgFile load(Long id) throws Exception;
 
 }
